refactor: separate LUT and CLUT into separate modules
